+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Make it possible to select USB device with serial number
+- A121: Vibration: Peak detection
 
 ### Fixed
 - A121 presence: Fix depth filter length so it never is bigger than